revert(switcher): removed the main switch due to QTimer errors
I suspect that the issue is this - the switcher is not the problem. The problem is
that if we populate the model and then show the QTableView, showing the view triggers
more work to the art_model.Model.data() method which can cause some side-effects for
threading.

So in short, if the view is already visible by the time
`_MaskedDataProxy.populate_rows` is called, everything is great. But if
it happens after or as a side-effect of the row population completing
- that causes issues.

The issue could be solved possibly by delaying the view from showing but
since this feature was nice-to-have, I decided it was less work to just
omit the main switcher.